[Decomp][GameObject] Cohesion split: extract Update + Use to GameObject*.cpp
